:recycle: Split file
main execut & main parsing
