+++ minishell/Includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tonted <tonted@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/11 18:35:40 by mdoquocb          #+#    #+#             */
-/*   Updated: 2022/09/16 19:32:14 by tonted           ###   ########.fr       */
+/*   Updated: 2022/09/27 13:31:47 by tonted           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,8 @@
 # include <fcntl.h>
 # include <stdbool.h>
 # include <errno.h>
+
+# define PROMPT "minishell %>"
 
 extern int	errno;
 
@@ -136,6 +138,12 @@ char	**tabstrdup(char **tabstr);
 
 /* utils */
 int	exit_free_envp(char ***envp);
+
+/* execute */
+int	execute(char *command, int *status, char ***envp);
+
+/* parsing */
+t_cmd	*parsing(char *command, int *status, t_cmd **cmd, char **envp);
 
 
 	//4_utils.c
